Moved DashboardTest to ui package, fix invisible coach names on map in dark mode
