+++ app/src/androidTest/java/com/github/sdpcoachme/ui/DashboardTest.kt
@@ -1,4 +1,4 @@
-package com.github.sdpcoachme
+package com.github.sdpcoachme.ui
 
 import android.Manifest.permission.ACCESS_FINE_LOCATION
 import android.content.Intent
@@ -15,6 +15,8 @@ import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.rule.GrantPermissionRule
+import com.github.sdpcoachme.BuildConfig
+import com.github.sdpcoachme.CoachMeApplication
 import com.github.sdpcoachme.ui.Dashboard.TestTags.Buttons.Companion.COACHES_LIST
 import com.github.sdpcoachme.ui.Dashboard.TestTags.Buttons.Companion.HAMBURGER_MENU
 import com.github.sdpcoachme.ui.Dashboard.TestTags.Buttons.Companion.LOGOUT
